메인 카테고리 선택 - 홈, 웹툰 ... 메인카테고리 클릭 시 노란 선 표시 - 웹툰 제외 임시데이터 표시(이벤트 등록)
